ci: split integration test in two due to limited resources
Github runners have limited resources, which may cause bottlenecks and
even errors when running integration tests on demanding charms. Splitting
the single integration test job into two prevent this from happening.
The new jobs are:
* Integration
* Observability Integration
