+++ .github/workflows/integration.yaml
@@ -44,9 +44,9 @@ jobs:
       - run: sudo apt update && sudo apt install tox
       - run: tox -e ${{ matrix.charm }}-unit
 
-  integration:
+  integration-charm-deployment:
     name: Integration Test
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-20.04
 
     steps:
       - name: Check out repo
@@ -56,7 +56,7 @@ jobs:
         uses: charmed-kubernetes/actions-operator@main
         with:
           provider: microk8s
-          channel: 1.21/stable
+          channel: 1.22/stable
           # Pinned to 2.3.34 due to https://bugs.launchpad.net/juju/+bug/1992833
           # This should be fixed on release of juju 2.3.36 + libjuju 3.0.3
           bootstrap-options: --agent-version="2.9.34"
@@ -68,6 +68,36 @@ jobs:
         run: tox -vve integration -- --model knative-test --destructive-mode
         env:
           KUBECONFIG: /home/runner/.kube/config
+
+  observability-integration:
+    name: Observability Integration Test
+    runs-on: ubuntu-20.04
+
+    steps:
+      - name: Check out repo
+        uses: actions/checkout@v2
+
+      - name: Setup operator environment
+        uses: charmed-kubernetes/actions-operator@main
+        with:
+          provider: microk8s
+          channel: 1.22/stable
+          # Pinned to 2.3.34 due to https://bugs.launchpad.net/juju/+bug/1992833
+          # This should be fixed on release of juju 2.3.36 + libjuju 3.0.3
+          bootstrap-options: --agent-version="2.9.34"
+      - run: juju add-model cos-test
+
+      - name: Run integration tests
+
+        run: tox -vve cos-integration -- --model cos-test --destructive-mode
+        env:
+          KUBECONFIG: /home/runner/.kube/config
+
+      - run: kubectl get pod/prometheus-k8s-0 -n knative-test -o=jsonpath='{.status}'
+        if: failure()
+
+      - run: kubectl get pod/knative-operator-0 -nknative-test -o=jsonpath='{.status}'
+        if: failure()
 
       - run: kubectl get all -A
         if: failure()
