+++ tests/test_bundle.py
@@ -6,7 +6,6 @@ import logging
 
 import pytest
 import requests
-import tenacity
 import yaml
 from pytest_operator.plugin import OpsTest
 
@@ -156,62 +155,3 @@ async def test_cloud_events_player_example(ops_test: OpsTest):
     assert post_req.status_code == 202
     get_req = requests.get(f"{url}/messages", allow_redirects=False, verify=False)
     assert get_req.status_code == 200
-
-
-# knative-operator is the charm that actually talks to prometheus
-# to configure the OpenTelemetry collector to be scraped
-APP_NAME = "knative-operator"
-
-
-async def test_prometheus_grafana_integration(ops_test: OpsTest):
-    """Deploy prometheus and required relations, then test the metrics."""
-    prometheus = "prometheus-k8s"
-    prometheus_scrape = "prometheus-scrape-config-k8s"
-    scrape_config = {"scrape_interval": "30s"}
-
-    # Deploy and relate prometheus
-    await ops_test.model.deploy(prometheus, channel="latest/beta", trust=True)
-    await ops_test.model.deploy(prometheus_scrape, channel="latest/beta", config=scrape_config)
-
-    await ops_test.model.add_relation(APP_NAME, prometheus_scrape)
-    await ops_test.model.add_relation(
-        f"{APP_NAME}:otel-collector", "knative-eventing:otel-collector"
-    )
-    await ops_test.model.add_relation(
-        f"{APP_NAME}:otel-collector", "knative-serving:otel-collector"
-    )
-    await ops_test.model.add_relation(
-        f"{prometheus}:metrics-endpoint", f"{prometheus_scrape}:metrics-endpoint"
-    )
-
-    await ops_test.model.wait_for_idle(status="active", timeout=60 * 20)
-
-    status = await ops_test.model.get_status()
-    prometheus_unit_ip = status["applications"][prometheus]["units"][f"{prometheus}/0"]["address"]
-    log.info(f"Prometheus available at http://{prometheus_unit_ip}:9090")
-
-    for attempt in retry_for_5_attempts:
-        log.info(
-            f"Testing prometheus deployment (attempt " f"{attempt.retry_state.attempt_number})"
-        )
-        with attempt:
-            r = requests.get(
-                f"http://{prometheus_unit_ip}:9090/api/v1/query?"
-                f'query=up{{juju_application="{APP_NAME}"}}'
-            )
-            response = json.loads(r.content.decode("utf-8"))
-            response_status = response["status"]
-            log.info(f"Response status is {response_status}")
-            assert response_status == "success"
-
-            response_metric = response["data"]["result"][0]["metric"]
-            assert response_metric["juju_application"] == APP_NAME
-            assert response_metric["juju_model"] == ops_test.model_name
-
-
-# Helper to retry calling a function over 30 seconds or 5 attempts
-retry_for_5_attempts = tenacity.Retrying(
-    stop=(tenacity.stop_after_attempt(5) | tenacity.stop_after_delay(30)),
-    wait=tenacity.wait_exponential(multiplier=1, min=1, max=10),
-    reraise=True,
-)
