+++ tox.ini
@@ -26,4 +26,11 @@ allowlist_externals = rm
 deps =
     -rtest-requirements.txt
 commands =
-  pytest --show-capture=no --log-cli-level=INFO -vvs --tb=native tests/ {posargs}
+  pytest --show-capture=no --log-cli-level=INFO -vvs --tb=native {posargs} tests/test_bundle.py
+
+[testenv:cos-integration]
+allowlist_externals = rm
+deps =
+    -rtest-requirements.txt
+commands =
+  pytest --show-capture=no --log-cli-level=INFO -vvs --tb=native {posargs} tests/test_cos_integration.py
